# docs/swagger.yaml
basePath: /
definitions:
  social-platform_internal_dto.CreatePostRequest:
    properties:
      content:
        maxLength: 5000
        type: string
    required:
    - content
    type: object
  social-platform_internal_dto.CreateUserRequest:
    properties:
      date_of_birth:
        type: string
      email:
        maxLength: 254
        type: string
      full_name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  social-platform_internal_dto.LoginRequest:
    properties:
      email:
        maxLength: 254
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  social-platform_internal_dto.PostLikeResponse:
    properties:
      created_at:
        type: string
      post_id:
        type: integer
      user_id:
        type: string
    type: object
  social-platform_internal_dto.PostResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      post_id:
        type: integer
      user_id:
        type: string
    type: object
  social-platform_internal_dto.UpdatePostRequest:
    properties:
      content:
        maxLength: 5000
        type: string
    type: object
  social-platform_internal_dto.UpdateUserRequest:
    properties:
      date_of_birth:
        type: string
      full_name:
        maxLength: 100
        type: string
    type: object
  social-platform_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Social Platform Backend APIs
  title: Social Platform APIs
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and set access and refresh tokens in HttpOnly
        cookies
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/social-platform_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.UserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Revoke the current refresh token and clear authentication cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User logout
      tags:
      - Auth
  /auth/me:
    get:
      description: Get information of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Rotate the refresh token and issue a new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /posts:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Create post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/social-platform_internal_dto.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/social-platform_internal_dto.PostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create post
      tags:
      - Posts
  /posts/{post_id}:
    delete:
      description: Delete a post. Only the post owner can delete it.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Delete post
      tags:
      - Posts
    get:
      description: Get a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.PostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get post by ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update a post. Only the post owner can update it.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Update post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/social-platform_internal_dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.PostResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update post
      tags:
      - Posts
  /posts/{post_id}/like:
    delete:
      description: Remove user's like from a post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Like not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Unlike a post
      tags:
      - Post Likes
    post:
      description: Like a post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/social-platform_internal_dto.PostLikeResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Like a post
      tags:
      - Post Likes
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/social-platform_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/social-platform_internal_dto.UserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create user
      tags:
      - Users
  /users/{user_id}:
    get:
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: User information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/social-platform_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/social-platform_internal_dto.UserResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - Users
  /users/{user_id}/follow:
    delete:
      description: Unfollow another user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully unfollowed
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Follow relationship not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Unfollow a user
      tags:
      - Follow
    post:
      description: Follow another user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Already following user
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Follow a user
      tags:
      - Follow
  /users/{user_id}/posts:
    get:
      description: Get posts created by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Number of posts
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/social-platform_internal_dto.PostResponse'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get posts by user
      tags:
      - Posts
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"